requirements.txt
gradio
openai-whisper
transformers
accelerate
bitsandbytes
gtts
langchain
langchain-community
langchain-core
huggingface_hub
numpy
torch
scipy
